refactor(version): Extract .promptVersion() -> lib/prompt-version
